Se creo la clase ciudad y sus test

+++ Aeropuerto/src/test/java/ar/edu/undec/AerolineaUnitTest.java
@@ -9,25 +9,25 @@ public class AerolineaUnitTest {
 	@Test
 	public void intanciarAerolinea_TodoslosAtributos_Correctos() {
 		Aerolinea unaAerolinea = new Aerolinea(1,"Avianca");
-		assertEquals (1, unaAerolinea.getIdAerolinea().intValue());
+		assertEquals (1, unaAerolinea.getIdAerolinea());
 		assertEquals ("Avianca", unaAerolinea.getNombre());
 	}
 	@Test
 	public void intanciarAerolinea_TodosLosAtributos_IdIncorrexto() {
 		Aerolinea unaAerolinea = new Aerolinea(1,"Avianca");
 		assertEquals("Avianca", unaAerolinea.getNombre());
-		assertNotEquals(0, unaAerolinea.getIdAerolinea().intValue());
+		assertNotEquals(0, unaAerolinea.getIdAerolinea());
 	}
 	@Test
 	public void intanciarAerolinea_TodosLosAtributos_NombreIncorrexto() {
 		Aerolinea unaAerolinea = new Aerolinea(1,"Avianca");
 		assertNotEquals("Aerolineas AR", unaAerolinea.getNombre());
-		assertEquals(1, unaAerolinea.getIdAerolinea().intValue());
+		assertEquals(1, unaAerolinea.getIdAerolinea());
 	}
 	@Test
 	public void intanciarAerolinea_TodosLosAtributos_Incorrectos() {
 		Aerolinea unaAerolinea = new Aerolinea(1,"Avianca");
 		assertNotEquals("Aerolineas AR", unaAerolinea.getNombre());
-		assertNotEquals(0, unaAerolinea.getIdAerolinea().intValue());
+		assertNotEquals(0, unaAerolinea.getIdAerolinea());
 	}
 }
